+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update identity proof validation API to use `proof_type` parameter.
 - Ignore `Announce(Add)` and `Announce(Remove)` activities from Lemmy.
 - Check uniqueness of issuers when saving identity proofs.
+- Check uniqueness of chain IDs when saving payment options.
 
 ### Fixed